test: test gamma_shift
