+++ tests/test_s11.py
@@ -1,15 +1,8 @@
-import pytest
+import numpy as np
 
-from edges_io.io import SwitchingState
-
-from edges_cal import s11_correction as s11
+from edges_cal import reflection_coefficient as rc
 
 
-@pytest.fixture(scope="module")
-def internal_switch(cal_data) -> SwitchingState:
-    return SwitchingState(cal_data / "S11" / "SwitchingState01")
-
-
-def test_bad_nterms(internal_switch: SwitchingState) -> None:
-    with pytest.raises(ValueError, match="n_terms must be >0"):
-        s11.InternalSwitch(internal_switch, n_terms=(1, 1, 0))
+def test_gamma_shift_zero():
+    s11 = np.random.normal(size=100)
+    assert np.all(s11 == rc.gamma_shifted(s11, 0, 0, 0))
